Automated_Resume_Parser/requirements.txt
Flask==3.0.0
python-docx
pdfplumber
spacy
psycopg2-binary
pandas